fix_1

+++ tests/archer_test.h
@@ -46,7 +46,7 @@ TEST(arrow_of_archer, shots){
         p.enemy_archer(time, 0);
     }
 
-    ASSERT_EQ(p.index_arrows.size(), 2);
+    ASSERT_EQ((int)p.index_arrows.size(), 2);
     ASSERT_EQ(alive, true);
 
 }
